+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.0.4
+
+🆕 New features:
+
+- Made service creation optional, added a no service example
+
+  ([PR #4](https://github.com/srb3/terraform-kubernetes-simple-deployment/pull/4))
+
+## 0.0.3
+
+🆕 New features:
+
+- Added option to pass labels to the deployment
+
+  ([PR #3](https://github.com/srb3/terraform-kubernetes-simple-deployment/pull/3))
+
 ## 0.0.2
 
 🔧 Fixes: